Add files via upload
The purpose of me creating this project, was to get a better sense and understanding of using networking in your code. I thank Tech With Tim for breaking down the concepts so well, you can really build onto this project to create a effective chatroom for a group of friends tht looks dope too. I'm not sure if ill build on this, but i feel like i should and see how interesting i can make it, im sure building on it will teach me even more about hosting a server.